+++ lib/lxp/packet/write_single.rb
@@ -27,14 +27,10 @@ class LXP
       #
       # Raise if not, as that is not expected?
       #
-      # Base#value will return an int for protocol 1, or an Array
-      # for protocol 2. If we can, convert that Array to an int.
-      #
       def value
         raise 'value_length not 2?' unless value_length == 2
 
-        r = super
-        r.is_a?(Array) ? Utils.int(r, 2) : r
+        Utils.int(values[0, 2])
       end
     end
   end